requirements.txt
asgiref==3.7.2
Django==4.2.8
sqlparse==0.4.4
typing_extensions==4.8.0

pip~=23.3.1
attrs~=23.1.0
Pillow~=10.1.0
Jinja2~=3.1.2
filelock==3.13.1
pytz~=2023.3.post1
setuptools==69.0.2
PyYAML~=6.0.1
numpy==1.24.4
requests~=2.31.0
packaging~=23.2
Cython==3.0.6
platformdirs==4.1.0
djangorestframework==3.14.0
coreapi~=2.3.3
uritemplate~=4.1.1
coreschema~=0.0.4
PyJWT==2.8.0